Admin view of prompts is now displayed in accordion view.  Built on top of Bootstrap Tables code.
My code is sloppy and amateur. There is probably much better praxis for what I'm doing and I am very open to any suggestions, pull requests, Issues, etc to try to help make the code better.
To make up for it, there is also a link to essentially the original prompts code.